Extends package refactor
